refactor(tracer): keep CLS CoreTracer the default and introduce CoreTraceBase
