+++ packages/opencensus-nodejs/src/trace/tracing.ts
@@ -41,7 +41,7 @@ export class Tracing implements core.Tracing {
 
   /** Constructs a new TracingImpl instance. */
   constructor() {
-    this.tracer = new core.CoreTracerCls();
+    this.tracer = new core.CoreTracer();
     this.defaultPlugins = PluginLoader.defaultPluginsFromArray(
         Constants.DEFAULT_INSTRUMENTATION_MODULES);
   }
