+++ packages/opencensus-nodejs/test/test-tracing.ts
@@ -225,7 +225,7 @@ describe('Tracing', () => {
       const tracing = new Tracing();
       tracing.start();
       const tracer = tracing.tracer;
-      assert.ok(tracer instanceof core.CoreTracerCls);
+      assert.ok(tracer instanceof core.CoreTracer);
     });
   });
